refactor: move VarInt reading/writing to convenience functions on Ser trait
